Bug#20035071: Out of range error in subselect lead to assertion failed
There is a missing error check during preparation of this query,
so that we reach the regular function return in
st_select_lex::prepare(). However, there has been an error when
evaluating the LIKE predicate, which has gone unnoticed.

The fix applies to Item_func_like::fix_fields(), check if the call
to val_str() returns an error. However, the real fix should be to
check for error in Item_real_func::val_str(). But since this function
has no easily accessible THD reference, it's better to do it this way.