BUG#20550991 - MAIN.DAEMONIZE_OPT FAILS IN VALGRIND
main.daemonize_opt fails in valgrind and the test is supposed
to be skipped under valgrind.
Fix MTR to use --leak-check=no so that valgrind errors are not
reported when server is started with --daemonize option.
Also check for invalid file descriptor before reading from pid file
and closing it.