Bug #20595292 	INNODB: MTR TEST 16K FAILING SPORADICALLY
The MTR test '16k' is failing on weekly-trunk is about 4 different ways.
The test needs to be made more deterministic.  The actual index ID, table ID
or space ID does not matter if there is any likelihood that it might be
different when this test is run on a database that had previously run various
other tests.

Change the I_S queries in this test, as well as similar queries in 8k.test
and 4k.test so that the results are predictable.

In addition, it has been observed that some test fail when they query
I_S.INNODB_SYS_DATAFILES where path like 'test' because the word
test can sometimes occur earlier in the path, depending on the system
it runs on and the test parameters such as embedded.  So every innodb
test that does that query is replaced with:
--source suite/innodb/include/show_i_s_tablespaces.inc
and when they also query SYS_TABLES, this test is also used instead:
--source suite/innodb/include/show_i_s_tables.inc