BUG#20031475: ASSERTION `HASH_NOT_OK == 0' FAILED IN               SQL_ACL.CC:1944
Description:- Improper deletion of users from mysql.user
table followed by granting privileges to a new user with
same name as before leads to server crash.

Analysis:- Deletion of users by deleting rows from the
mysql.user table and granting privileges to a new user with
same name as before leads to server crash. This is due to
the lack of reloading acl privileges. A manual FLUSH
PRIVILEGES command after the improper deletion will fix this
problem. The assertion failure which occurs during the
"acl_update()" is also due to the incorrect length of the
plugin passed to the "acl_update_user()". Fixing this
problem solves the server crash.

Fix:- Correct plugin length is passed to "acl_update_user()".

NOTE:- This patch is only for mysql-5.6. In mysql-5.7, the
bug does not exist anymore.