WL#7703 InnoDB: Change defaults and deprecate settings in next 5.7 release
InnoDB will change the default values of the following settings
in the next 5.7 release:

innodb_strict_mode=ON (was: OFF)

innodb_large_prefix=ON (was: OFF)
innodb_file_format=Barracuda (was: Antelope)

We will also deprecate the following 4 parameters:

innodb_large_prefix
innodb_file_format
innodb_file_format_check
innodb_file_format_max

If non-default startup values are used for the 4 deprecated
parameters, InnoDB will issue a warning to the server error log.

We return the same deprecation and removal warning to the client when
any of the 4 deprecated parameters are set by the user dynamically in
the running server using a SET statement.

The variable innodb_strict_mode will not be deprecated, and no warning
will be issued when it is used with the SET statement.

The bulk of the changes is due to removing assignments to
innodb_file_format=Barracuda in many tests.

All code changes are to ha_innodb.cc. Here is a summary:

innobase_init(): Issue a deprecation warning to the server error log
if innodb_file_format or innodb_file_format_max is specified in the
configuration, or if innodb_file_format_check=OFF or innodb_large_prefix=OFF.
(The Boolean parameters will default to ON.)

innodb_file_format_name_update(): Issue a warning to the client.
This is invoked by SET GLOBAL innodb_file_format.

innodb_file_format_max_update(): Issue a warning to the client.
This is invoked by SET GLOBAL innodb_file_format_max.

[New] innodb_large_prefix_update(): Issue a warning to the client.
This is invoked by SET GLOBAL innodb_large_prefix.

RB: 7778
Reviewed-by: Vasil Dimov <vasil.dimov@oracle.com>
Reviewed-by: Annamalai Gurusami <annamalai.gurusami@oracle.com>