BUG#20528928 - BUFFER OVERFLOW IN MYSQL_QUERY_CACHE_INVALIDATE4
Fix and Description:
The function mysql_query_cache_invalidate4(), which is
invoked by InnoDB when executing ON DELETE or ON UPDATE
clauses in FOREIGN KEY constraints, is using an
insufficiently dimensioned array for a filename-safe
encoded database(schema) name.

The fix is to correct size of the array to hold the
max possible length of the encoded database (schema) name.
In addition, the patch also corrects the comment in
the function mysql_query_cache_invalidate4().