BUG#20492319: LOG LOCK MAY NOT BE UNLOCKED IF ADD_LOGGED_GTID FAILED.
When a Gtid_log_event is received from the master,
and after being queued in the relay_log, its GTID
is saved in the received GTID set. This happens
inside queue_event and is done while holding a lock,
the relay_log lock. The function that is called to
save the GTID can return an error. That is fine
and is being tested properly. However, in the event
of an error there will be a jump to the end of the
queue_event function, but the relay_log mutex ends
up NOT being released.

This patch adds the missing mysql_mutex_unlock call.