Bug#20460208: !table || (!table->read_set || bitmap_is_set)
The optimizer relies on thd->mark_used_columns to have the value
MARK_COLUMNS_READ for columns in WHERE conditions, etc.
If this is not the case, fields are sometimes erroneously marked for
write or not marked at all. In particular, if a failed statement of
some "wrong" type preceded a correct statement, this problem might hit.
In the bug report, a failed INSERT statement is followed by a DELETE
statement.

Followup patch: Make sure that we start all executions of prepared
statements too with thd->mark_used_columns = MARK_COLUMNS_READ.