BUG#20629125: MAKE RPL.RPL_SP_PRIVILEGES TEST WORK WITH GROUP REPLICATION
rpl.rpl_sp_privileges test doesn't use tables with primary keys, a
requirement of Group Replication, apart from that all test is
supported so we should had primary keys to make the test run with
Group Replication.

To solve the above issue primary keys were added to the tables of
the test.