Bug#20454833 PREFIX INDEX FIELD, ON AN INAPPROPRIATE DATA TYPE Bug#20506527 FAILING ASSERTION: !MBMAXLEN || !(PREFIX_LEN % MBMAXLEN)
Both bugs have the same cause; I'll describe the scenario of the first
one. Query is:
select `t2`.`b`,`t1`.`b`,`t1`.`a`
from `t2` inner join `t1` on 1 cross join `t2` `a`  on 1
group by `t1`.`b`;

In trunk as in 5.6, optimizer decides to create a tmp table
for the grouping. It will have 3 columns: t2.b (blob), t1.b (int),
t1.a (blob). Notice the same name (but different source table) for
the two first.
An index is wanted, on t1.b (for grouping on it).
In 5.6, it works.
In trunk (if internal_tmp_disk_storage_engine=innodb), we go into
ha_innodb.cc's create_index(), which has this loop to find which Field
is related to the key_part (key_part describes the index which the
optimizer asks to create):

     for (ulint i = 0; i < key->user_defined_key_parts; i++) {
         KEY_PART_INFO*    key_part = key->key_part + i;
...
         Field*    field = NULL;
         for (ulint j = 0; j < form->s->fields; j++) {

             field = form->field[j];

             if (0 == innobase_strcasecmp(
                     field->field_name,
                     key_part->field->field_name)) {
                 /* Found the corresponding column */

                 goto found;

See how it compares field names. key_part->field->field_name is "b" (t1.b);
in our case, t2.b is the first column of form->s->fields, its field_name
is "b", so we have a (wrong) match; then things go wrong as expected
(field t2.b is a blob, though the index was supposed to be on an int (t1.b)),
and we have:
[ERROR] MySQL is trying to create a column prefix index field, on an
inappropriate data type. Table name tmp/#sql_4d6e_0, column name b.

The loop above was surely relying on the fact that in a user-created
table, field names are unique. But this is not (yet) true for internal
tmp tables.
Fix: replace the loop with:
field=form->field[key_part->field->field_index];
I verified with dbug_assert + running mtr, that the loop above effectively
finds "field" equal to the proposed form->field[etc], except in the bugs'
testcases.