Fix for bug#20538179
 	NDB : TEST HANDLING OF ARBITRATOR FAILURE DURING ARBITRATION

Fix a failure seen in AutoTest: 'testNodeRestart -n ClusterSplitLatency'.
This is also a genuine bug where the 'Loser' in the arbitration
is allowed to reconnect before the arbitration has completed.
During the reconnect, it hits a ndbrequire in Qmgr::execRECONNECT_REP
as we are still in phase ZRUNNING.

Frazer describes the failure scenario as:

.........
I guess that it needs :
 1.  Losing side is slightly slower / behind the winning side at detecting the silence / loss of the other side
 2.  Losing side enters state ZPREPARE_FAIL as part of execPREP_FAILREQ() - in 2 node case, this occurs quickly once the 'other node' failure is decided, as we are / must become President. Losing side sends CLOSE_COMREQ to TRPMAN for 'other' node.
 3.  Handling of 'other node' failure causes DUMP 9991 to be sent to TRPMAN, re-enabling communication.
 4.  Unblocking of transporter causes DISCONNECT_REP to be sent to TRPMAN and then QMGR
 5.  DISCONNECT_REP reception calls node_failed() which causes transition from ZPREPARE_FAIL -> ZFAIL_CLOSING due to bad (fixed) switch statement
      Also causes all pending messages to be discarded, so e.g. no FAIL_REP signals are received which would kill us.
 6.  Losing node is in state CHOOSE_ARBIT, waiting for response.
 7.  Arbitrator disconnects - no action yet, still waiting for ArbitTimeout
 8.  As we're in ZFAIL_CLOSING, the 'other side' can disconnect and reconnect
 9.  If the 'other side' has finished handling our failure, and attempts to reconnect then we can generate CONNECT_REP resulting in ndbrequire(false) etc...

So I think it's a genuine bug, but we probably don't often see it as it requires the 'DISCONNECT_REP' to overtake the 'FAIL_REP' signals in the transporter from the 'other' node.  The error insertion / dump codes probably make that more likely.

.........

The fix is to change Qmgr::node_failed(), such that when in phase
PREPARE_FAIL it does nothing, instead of sending another (redundant)
CLOSE_COMREQ, which would have replaced the PREPARE_FAIL state with
ZFAIL_CLOSING, which would have allowed reconnection after the
3 second grace period.