BUG#20504513 VALGRIND ERROR IF IPV6 IS NOT SUPPORTED
Description and Fix:
If IPv6 is not supported, an error reported in error
log when using valgrind.

In network_init function, we create a dummy IPv6 socket.
If IPv6 is enabled a valid socket file descriptor is
returned else -1 is returned. Irrespective of the return
value, we close the socket and hence there is a valgrind
error happens when IPv6 not enabled. The fix is close
the socket file descriptor only if the descriptor is
valid. This patch is based on a contribution from
Zhai Weixiang via OCA.