WL#8016, post-push fix and cleanup
Before this fix the hint parser silently ignored the rest of query text
in the case of open hint commentary (see the "error" branch of the 1st
rule in the hint parser grammar). The main parser usually failed with a
syntax error then, since the query text was truncated after SELECT,
INSERT, REPLACE, UPDATE or DELETE keyword, so the problem was masked.

The current fix removes the "error" branch of the 1st hint parser grammar
rule: that forces hint parser to abort on open commentaries and return
error to the main parser.

Also the patch updates the hint parser to output same error messages for
open hint commentaries like we do for regular ones:
  "... near '/*...' at line N"
instead of
  "... near '' at line M".