Bug#20527217 - INNODB: PUT FTS AUXILIARY TABLES INTO THE SAME GENERAL TABLESPACE AS THE PRIMARY
Problem:
--------
If an FTS primary table is assigned to a general tablespace, FTS
auxiliary tables should also be created in the same tablespace.

Fix:
----
Create FTS aux tables in same general tablespace, if the FTS primary
table is in general tablespace.

Since FTS Aux tables will be in the same general tablespace, they follow
the  properties of the tablespace.

1) AUX Tables can be now compressed (if General Tablespace is compressed)
2) AUX Tables will have same ROW_FORMAT as FTS primary table
3) AUX Tables will now adhere to 'DATA DIRECTORY'

Reviewed-by: kevin.lewis@oracle.com
Reviewed-by: jimmy.yang@oracle.com
RB: 8166