Bug #18279587 FTS: ANOTHER CRASH IN HA_MYISAM::FT_INIT_EXT
Crash after:
drop table if exists t1;
create table t1 (aint) engine=myisam;
delete from t1 where a <> 1 order by match(a) against ('+aaaa' in
boolean mode);

Analysis:
This bug started to appear after bug fix for Bug #16073689: CRASH IN
ITEM_FUNC_MATCH::INIT_SEARCH was applied. This bug fix was applied in
versions: 5.1, 5.5, 5.6, 5.7 which explains why we get the new crash in
those versions.
The bug fix for Bug #16073689 adds a return statement (to function
fix_index) when fixed flag is not true. However, at the end of fix_index
function we have verification if we allow search on no indexed columns
(which is true for boolean search on MyISAM table). With this new return
statement this verification was skipped.
Bug fix:
Added verification if we allow search on no indexed column if fixed flag
is not true.