Bug#20487336: Assert 'length > 0 && keyparts != 0' in calc_length_and_keyparts
The real fix for the problem.

The problem was that fix_tables_after_pullout() was too eager to
adjust TABLE_LIST::sj_depends_on and sj_corr_tables for a semi-join:
It was called for semi-join nests that had already been adjusted
when merging a derived table/view.

The current solution calls fix_tables_after_pullout() only for
the derived table/view that is merged. Previously, this function
was called for all tables of a query block.

This solution will only work if derived tables are merged in a strict order,
see the comment above fix_tables_after_pullout() for details.

It was also necessary to convert fix_tables_after_pullout() to be called
for a join nest instead of a list of join nests.

The changed plan for the query in derived.test was due to there being
a semi-join inside a derived table, and the sj_depends_on field was
adjusted wrongly due to this bug.