Bug#17435114 ASSERTION FAILED: THD->LEX->LOCK_TABLES_STATE == QUERY_TABLES_LIST::LTS_NOT_LOCK
Adding testcase.
Bug was introduced in 07/2013 and can't be reproduced neither with a
tree of 09/2013 nor today.
It was previously debugged, and found to be due to two lock_tables()
calls. Debugging today's 5.7 shows only one lock_tables() call, which
is correct:
- no need to lock subquery's tables as ALTER EVENT is designed
to reject subqueries which use tables, see 'case SQLCOM_ALTER_EVENT' in
mysql_execute_command()
- just need to lock mysql.events for update.
Likely, the many refactorings done in the last 1.5 years have
changed the code flow.