Bug#20559828 SHOW GLOBAL VARIABLES WHERE NOT DEPRECATED IN EMBEDDED
Before this fix, the command:
  SHOW GLOBAL VARIABLE WHERE ...
printed a deprecation warning, inviting to use the performance_schema tables
instead.

This warning was incorrect, and caused confusion for the embedded build.

There are two separate changes that are independent:

First, the syntax
  SHOW GLOBAL VARIABLE WHERE ... [LIKE ...]
should be changed to simply
  SHOW GLOBAL VARIABLE [like ...]
abandonning the WHERE clause syntax.
This applies to all builds, including embedded.

Secondly, the query
  SHOW GLOBAL VARIABLE [like ...]
might be rewritten using the new performance_schema tables instead.
This is optional, and only applies to builds that include the
performance_schema (hence does not apply for the embedded build).

The warning printed mixed the two changes.

The deprecation only applies to the first part,
dropping support for the WHERE clause in SHOW STATUS / SHOW VARIABLES.

With this fix, the proper deprecation warning is now printed,
and it is also printed for embedded builds.