Bug#20574628: SEMI-SYNC REPLICATION PERFORMANCE DEGRADES WITH A HIGH NUMBER OF THREADS Problem: when semi-sync replication is enabled and when the number of threads running is increasing (going beyond a threshold), performance of the system is going drastically down.
Analysis:
A transaction in semi-sync replication waits for the ACK
inside ReplSemiSyncMaster::commitTrx on a condition variable
COND_binlog_send_. In reportReplyBinlog function, when server
receives an ACK, it sends condition broadcast to all the
threads that are waiting on COND_binlog_send_.

Each transaction thread waiting an ACK does the follow (simplified):
- hold a lock LOCK_binlog_
- in a loop ( while (is_on()) ) :
  - compare the ACK position with its own position;
     - if the ACK is still behind, it waits on a condition variable COND_binlog_send_
     - if the ACK is ahead or equal, it exits this loop and
        proceeds with the next step
- unlock the lock LOCK_binlog_

The problem here with this design is that all the threads wake up and while it
is likely that only few will return to the application and most of them will
go back to wait on the same condition variable. This creates unnecessary context
switches and a lot of LLC cache thrashing in multi-cpu systems
which reduces the throughput of the system.

Fix: The approach splits the *one* condvar used for all threads
to wait into *several* condvars. Each of the condvars is associated
with a position range. Thence only those that fall threads that are waiting
for a given acknowledged range are awaken when the ACK comes back
(as opposed to waking up all transactions, regardless whether the ACK affects
them or not).