Bug#20652173 MDL FOR TABLESPACES MUST BE REFLECTED IN P_S
MDL for tablespaces was introduced without the required
updates to the P_S implementation. This bugfix adds the
required P_S changes to map from the new MDL namespace
to an entry in a P_S internal enumeration, mapping to
correct object type name, etc.

A new test case is added to the MDL tablespace test to
verify correct P_S output as well as proper synchronization
with I_S contents.