BUG#20562022: ED_CONNECTION::EXECUTE_DIRECT() USES UNINITIALIZED VALUES
Problem: Valgrind issues a warning about the variable
LEX::keep_diagnostics being read without prior initialization. Root cause
is that the LEX constructor does not initialize it and in certain
circumstances (NDB) it is read before normal initialization takes
place.

Solution: Introduce a new enum value DA_KEEP_UNSPECIFIED, and use this
to initialize the member variable in the constructor. This ensures
that it is always safe to read the member variable, but retains the
semantics of the existing initialization mechanism.