BUG#20545943 patch 1/3: queue_event
BUG#20545943: MY_MESSAGE() WITH WRONG NUMBER OF ARGUMENTS IN RPL_SLAVE.CC:9540

In this bug, my_message was used with the wrong number of arguments in
queue_event and start_slave(THD*,...).  The root cause was that these
functions have many error conditions, each generating different error
messages.  In each error condition it stored the ER_* constant in a
variable, and then it reported the error in a one centralized place
near the end of the function.

Since different error messages need different numbers of arguments, it
is hard to maintain such code.  Therefore we fix the bug by reporting
the errors exactly where they happen.  That creates less
dependencies between different parts of the code.

This also makes it unnecessary to return error codes between
functions, making the protocols for the functions simpler.

This bug is split into three patches:
 1. Fix queue_event.
 2. Make a small simplification in the prototype of stop_slave.
 3. Fix stop_slave.