Bug#19504183 GIS: -0 WOULD BE TRANSLATED TO 0 AUTOMATICALLY
Problem: The handling of negative zero in geometric points is
inconsistent. Most functions preserve negative zeros, but some
function convert negative zeros to zeros:

WKT export functions convert negative zero to zero explicitly in
Geometry::append_points(), Gis_point::get_data_as_wkt() and
Gis_line_string::get_data_as_wkt()

ST_AsGeoJSON() converts negative zeros to zeros if max_dec_digits is
larger than 308 (default is 2**32-1). This happens because
Item_func_as_geojson::append_coordinates() calls my_double_round(),
which makes an internal calculation where it tries to multiply the
value with 10**max_dec_digits. Non-zero values will result in +/-Inf,
but zero and negative zero will result in -NaN. Later in the code,
only the +/-Inf case is handled, so the function returns -NaN for zero
and negative zero. This is later converted to zero in GeoJSON output.

ST_GeomFromGeoJSON() uses rapidjson to parse the JSON string. "-0.0"
is handled correctly, but "-0" is stored as integer internally and
hence loses the sign bit.

Negative zero and zero have different binary representations, but they
compare as equal, so it's mostly a presentation issue.

Fix: Always preserve negative zeros.

In WKT export functions: Partially revert fix for bug #19133043, which
introcued negative zero to zero conversion.

In my_double_round(), add correct handling of zero and negative zero
for large numbers of decimals.

In rapidjson::GenericReader::ParseNumber(), store "-0" as a double
internally, but add a new parameter is_int to the Double() handler
function to make it possible to tell that the number was specified as
an integer.

Because of the storage format of geometric data and how the equality
operator works, geometries with negative zeros do not compare as equal
to the same geometries with zeros. This bugfix doesn't change
that. Geometric operations should be compared using ST_Equals instead
of the equality operator.