Bug#20691920 SOMETHING IN TEST OPT_HINTS CAUSES ALL MEMORY TO BE CONSUMED AFTER FIX 20685187
Regression introduced by:
Bug#20685187 USE -STD=C99 WHEN BUILDING WITH GCC ON SOLARIS

Fix: ensure that support for timers is compiled in for Solaris.
Move opt_hint timeout tests to separate file
in case we have platforms that do not support timers.

Rollback the use of -std=c99
Add extra cmake check that isinf is visible in C and C++