WL#8374: Group Replication: plugin rebranding
This patch tries to accomplish several tasks.
Rename code and test files to remove the 'gcs' keyword.
Refactor the code and comments to remove the words "node" and
"cluster".
Refactor the code to remove the 'gcs' keyword from the code whenever
it did not make sense.
Update the README files.