Bug#20676000: FAILING MDL TABLESPACE ASSERTION FOR ALTER NDB TABLE IN UNIREG.CC:263
For mysql versions before 50120, NDB stored the tablespace names only
in the NDB dictionary. Thus, we have to get the tablespace name from
the engine rather than the .FRM file in this case.

This patch introduces a new handlerton function 'get_tablespace()', which
retrieves a tablespace name from a storage engine, given a schema- and table
name. The new handlerton function is implemented for the NDB storage engine.

The existing server function 'get_tablespace_name()', currently retrieving
the tablespace name from an .FRM file, is extended to instead get the name
from the storage engine, using the new handlerton function, in the cases
where this is relevant.