Bug #20536590: BACKPORT BUG #12368204 AND BUG #16953758 TO MYSQL-5.6
Backported the fixes.
Partially backported the fix for bug #19463877 too :
the part about different audit events count for the different
protocols.
Changed the strategy of the original fix to account for the lack
of the changes done in WL#6613 as follows:
Instead of mixing the calls to the audit log API by calling them
both outside of the general log hooks and inside some
moved all audit log API calls inside the general log hooks but
before any checks to the general log options. Made sure that the
general log hooks are called regardless of the value of the
general log option and moved all checking of it inside the
hooks.
Changed the binary log calles used by the prepared statement
code to use the convenience hooks instead of directly calling the
LOGGER functions. Ensured that prepared statements pass down the
query text.
Getting it from the THD is not correct for prepared statements.
** Added an extra test file that tests firewall with general log off.